requirements.txt
streamlit>=1.28.0
google-api-python-client>=2.100.0
serpapi>=0.1.5
openai>=0.28.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
fpdf2>=2.7.0
